Fixed position bug for teleport
Corrected the teleport placement, which was off by 0.5f resulting in the player being put into walls. Fixed a similar problem with the check of who should be teleported which was off by the same amount. Also, additional documentation, clean up, and reuse of an existing method (getting max floor height).