Don't show upper walls for skyboxes
There are a lot of weird corner cases, but I handled a few specifically for the skybox portion. The comments next to the extra checks describe what they do, but basically I extended the if check for showing wall tiles by adding a couple more conditions. I added another check for not showing the upper wall of a skybox when its ceiling was lower, but its floor was higher. This makes it consistent with the original code when the floor wasn't higher. You can always just make the sky's ceiling higher to avoid that behaviour, but it seemed better to be consistent.