Fix flatten tiles and control points not refreshing.
Two main fixes in this one: 1) The flatten tile wouldn't work with a negative selection width or height. A lot of other methods had if logic to handle, so I fixed at the source (where width and height are set) and then removed the logic in all of the other places. No need to remember to cut and paste that logic any more. 2) When selecting a new tile it would not always clear the control points from the old tiles. I added a control point clear on selection of a new tile, so now the control point will move to the newly selected tile right away (no need to re-click or toggle vertex mode or anything).