Fixed skybox placeholder texture showing for some levels
The win-screen-splash.bin and camp-ending.bin files had a few places where the skybox placeholder texture was being shown. This is because their (upper) walls were not marked as skybox, only their ceilings, and the empty blocks behind them had lower floors. Ideally, the tiles in the levels would have their walls set to skybox, but this will handle it too.